* commit '6067958aa1febfb34b392f0b902b90b4e95539d0':
  use longs instead of ints to store pointers in regex classes

+++ luni/src/main/java/java/util/regex/Pattern.java
@@ -285,7 +285,7 @@ public final class Pattern implements Serializable {
     private final String pattern;
     private final int flags;
 
-    transient int address;
+    transient long address;
 
     /**
      * Returns a {@link Matcher} for this pattern applied to the given {@code input}.
@@ -452,6 +452,6 @@ public final class Pattern implements Serializable {
         compile();
     }
 
-    private static native void closeImpl(int addr);
-    private static native int compileImpl(String regex, int flags);
+    private static native void closeImpl(long addr);
+    private static native long compileImpl(String regex, int flags);
 }
